Move AppState from presentation to domain layer
AppState is a domain model containing only domain concepts (AppMode,
Screen, TimeRangeTab, LocalDate) and computed properties with business
logic (isDemoMode, skipCredentialsCheck, currentTimeRangeTab). It has
no UI or framework dependencies, so it belongs in the domain layer.

Updated all imports across 10 files to reflect the new package location.
