+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/di/AppFeatureFactory.kt
@@ -1,10 +1,10 @@
 package space.be1ski.vibits.shared.app.di
 
 import kotlinx.datetime.LocalDate
+import space.be1ski.vibits.shared.app.domain.model.AppState
 import space.be1ski.vibits.shared.app.presentation.AppAction
 import space.be1ski.vibits.shared.app.presentation.AppEffect
 import space.be1ski.vibits.shared.app.presentation.AppEffectHandler
-import space.be1ski.vibits.shared.app.presentation.AppState
 import space.be1ski.vibits.shared.app.presentation.appReducer
 import space.be1ski.vibits.shared.core.elm.Feature
 import space.be1ski.vibits.shared.core.elm.FeatureImpl
