+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/domain/model/AppState.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.shared.app.presentation
+package space.be1ski.vibits.shared.app.domain.model
 
 import kotlinx.datetime.LocalDate
 import space.be1ski.vibits.shared.app.domain.model.Screen
